Migrate the new-issue submit route and split comment pagination out
Issue #407 merged a new GitHub route while this branch was open. Because
the unsafe helper was deleted rather than deprecated, that route failed to
compile in CI instead of silently shipping another worker that mutates
iocraft state off the render thread. Migrated it to spawn_gh_work.

Integrating that route pushed issues_dispatch.rs past the 850-line handler
boundary. Rather than relax the limit, moved the comment-pagination route
into its own module: it is a self-contained unit (eligibility, request-id
allocation, the gh call, and its typed events) that only shared two symbols
with the detail routes.
