+++ project-plans/issue437-plan.md
@@ -102,7 +102,8 @@ The all-routes option must replace this section with child-slice/stacked-PR cont
 | `catch_unwind` still runs the default panic hook | In-scope only for selected state-free recoverable request boundary | Required to meet the no-TUI-dump acceptance without hiding uncontained panics. A global arbitrary-panic ingestion queue remains out of scope. |
 | Current panic callbacks access the same copied iocraft state after a worker panic | Root-cause defect | Error-page recording cannot be reliable until the selected worker routes become state-free. |
 | Existing Errors capture is additive to inline mode errors | Accepted behavior change for worker panics only | Caught worker panic delivery must append directly and clear pending state without setting the inline mode error slot, matching the reporter's non-immediate preference. Ordinary `GhError` behavior remains unchanged. |
-| Remaining legacy routes outside the selected scope | Resolved by full migration | All 31 legacy production callers now use `spawn_gh_work`; `spawn_gh_task_with_panic` is deleted, so the old signature no longer exists to copy. |
+| A 32nd call site (`new_issue_submit.rs`, issue 407) merged to main while this branch was open | In-scope-Fix | Deleting the helper turned an otherwise silent semantic conflict into a CI compile error, which is the intended guardrail. The new route was migrated to `spawn_gh_work` during rebase. |
+| Remaining legacy routes outside the selected scope | Resolved by full migration | All legacy production callers (31 at branch point, 32 after integrating issue 407) now use `spawn_gh_work`; `spawn_gh_task_with_panic` is deleted, so the old signature no longer exists to copy. |
 | No deterministic way to trigger a real worker panic from the TUI | Rejected adding a production panic-injection hook | A `JEFE_*` panic-injection env var would be test-only behavior inside the shipped binary, which this plan lists as an explicit stop condition. The panic-to-Errors path is instead proven by `silent_route_panic_is_recorded_without_leaving_the_active_screen`, which drives a real panicking worker through the real delivery queue in an iocraft mock terminal and asserts both the recorded entry and the unchanged active screen. The real Errors screen is separately proven by the `errors-mode.json` TUI scenario. |
 
 ## Review counters
